Refresh participants list after adding one
